requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.2
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
redis==6.4.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
zope.interface==8.0.1
